Fixed race-condition in RTP SeqNum generation. Resolves #484. (#485)
* Fixed race-condition in RTP SeqNum generation. Resolves #484.

* Added safety valve to MediaStreamTrack.GetNextSeqNum(). Should never happen but only local variables, which should not really impact performance.

If this exception is ever thrown, something with the call-sites of MediaStreamTrack.GetNextSeqNum() is seriosly wrong. Many threads would have to simultaniously send RTP-Packages, which is most likely a bug. In normal operation, only one thread should create/send RTP-Packets and/or the time it takes to calculate the next SeqNum is really tiny.

Co-authored-by: Tobias Staehli <tobias.staehli@novalink.ch>