Move the AudioSink code into appropriate IAudioSink methods
In previous behaviour, all AudioSink initialization was made at same time as the AudioSource initialization. This cause issue if we create a RTPSession with only AudioSink (or with another AudioSource than the WindowsAudioEndpoint), by splitting the code we get more consistant behaviour. However this requires an update of the SIPSorcery VoipMediaSession object which shoudl call both IAudioSource and IAudioSink.  I submited the required change in that separate repository.